// src/open-api/model/currencySeparator.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type CurrencySeparator = ',' | '.';

export const CurrencySeparator = {
    Comma: ',' as CurrencySeparator,
    Period: '.' as CurrencySeparator
};