// src/open-api/model/currencySymbolPosition.ts
/**
 * Receipt Wrangler API.
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type CurrencySymbolPosition = 'START' | 'END';

export const CurrencySymbolPosition = {
    Start: 'START' as CurrencySymbolPosition,
    End: 'END' as CurrencySymbolPosition
};